requirements.txt
fastapi
uvicorn[standard]
gymnasium
pytest
httpx
pydantic
python-dotenv
anyio
torch
transformers
trl
wandb
openenv-core @ git+https://github.com/meta-pytorch/openenv.git
